+++ test/helpers/mutation-observer-mock.ts
@@ -1,3 +1,0 @@
-import 'mutationobserver-shim';
-
-global.MutationObserver = window.MutationObserver;