+++ tests/ids-axis-chart/ids-axis-chart.spec.ts
@@ -1,0 +1,39 @@
+import AxeBuilder from '@axe-core/playwright';
+import { expect } from '@playwright/test';
+import { test } from '../base-fixture';
+
+test.describe('IdsAxisChart tests', () => {
+  const url = '/ids-axis-chart/example.html';
+
+  test.beforeEach(async ({ page }) => {
+    await page.goto(url);
+  });
+
+  test.describe('general page checks', () => {
+    test('should have a title', async ({ page }) => {
+      await expect(page).toHaveTitle('IDS Axis Chart Component');
+    });
+
+    test('should not have errors', async ({ page, browserName }) => {
+      if (browserName === 'firefox') return;
+      let exceptions = null;
+      await page.on('pageerror', (error) => {
+        exceptions = error;
+      });
+
+      await page.goto(url);
+      await page.waitForLoadState();
+      await expect(exceptions).toBeNull();
+    });
+  });
+
+  test.describe('accessibility tests', () => {
+    test('should pass an Axe scan', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+      const accessibilityScanResults = await new AxeBuilder({ page } as any)
+        .exclude('[disabled]') // Disabled elements do not have to pass
+        .analyze();
+      expect(accessibilityScanResults.violations).toEqual([]);
+    });
+  });
+});