+++ tests/ids-tree/ids-tree.spec.ts
@@ -1,0 +1,66 @@
+import AxeBuilder from '@axe-core/playwright';
+import percySnapshot from '@percy/playwright';
+import { expect } from '@playwright/test';
+import { test } from '../base-fixture';
+
+import IdsTree from '../../src/components/ids-tree/ids-tree';
+
+test.describe('IdsTree tests', () => {
+  const url = '/ids-tree/example.html';
+
+  test.beforeEach(async ({ page }) => {
+    await page.goto(url);
+  });
+
+  test.describe('general page checks', () => {
+    test('should have a title', async ({ page }) => {
+      await expect(page).toHaveTitle('IDS Tree Component');
+    });
+
+    test('should not have errors', async ({ page, browserName }) => {
+      if (browserName === 'firefox') return;
+      let exceptions = null;
+      await page.on('pageerror', (error) => {
+        exceptions = error;
+      });
+
+      await page.goto(url);
+      await page.waitForLoadState();
+      await expect(exceptions).toBeNull();
+    });
+  });
+
+  test.describe('accessibility tests', () => {
+    test('should pass an Axe scan', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+      const accessibilityScanResults = await new AxeBuilder({ page } as any)
+        .exclude('[disabled]') // Disabled elements do not have to pass
+        .analyze();
+      expect(accessibilityScanResults.violations).toEqual([]);
+    });
+  });
+
+  test.describe('snapshot tests', () => {
+    test('should match innerHTML snapshot', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+      const handle = await page.$('ids-tree');
+      const html = await handle?.evaluate((el: IdsTree) => el?.outerHTML);
+      await expect(html).toMatchSnapshot('tree-html');
+    });
+
+    test('should match shadowRoot snapshot', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+      const handle = await page.$('ids-tree');
+      const html = await handle?.evaluate((el: IdsTree) => {
+        el?.shadowRoot?.querySelector('style')?.remove();
+        return el?.shadowRoot?.innerHTML;
+      });
+      await expect(html).toMatchSnapshot('tree-shadow');
+    });
+
+    test('should match the visual snapshot in percy', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+      await percySnapshot(page, 'ids-tree-light');
+    });
+  });
+});